+++ amethyst_renderer/src/renderer.rs
@@ -1,7 +1,7 @@
 use amethyst_error::Error;
 use fnv::FnvHashMap as HashMap;
 use gfx::memory::Pod;
-use winit::{dpi::LogicalSize, EventsLoop, Icon, Window as WinitWindow, WindowBuilder};
+use winit::{dpi::LogicalSize, EventsLoop, Window as WinitWindow, WindowBuilder};
 
 #[cfg(feature = "profiler")]
 use thread_profiler::profile_scope;
@@ -192,35 +192,8 @@ impl RendererBuilder {
 
     /// Applies configuration from `Config`
     pub fn with_config(&mut self, config: DisplayConfig) -> &mut Self {
-        self.config = config;
-        let mut wb = self.winit_builder.clone();
-        wb = wb
-            .with_title(self.config.title.clone())
-            .with_visibility(self.config.visibility);
-
-        if self.config.fullscreen {
-            wb = wb.with_fullscreen(Some(self.events.get_primary_monitor()));
-        }
-
-        let hidpi = self.events.get_primary_monitor().get_hidpi_factor();
-
-        if let Some(dimensions) = self.config.dimensions {
-            wb = wb.with_dimensions(LogicalSize::from_physical(dimensions, hidpi));
-        }
-
-        if let Some(dimensions) = self.config.min_dimensions {
-            wb = wb.with_min_dimensions(LogicalSize::from_physical(dimensions, hidpi));
-        }
-
-        if let Some(dimensions) = self.config.max_dimensions {
-            wb = wb.with_max_dimensions(LogicalSize::from_physical(dimensions, hidpi));
-        }
-
-        if let Some(ref icon) = self.config.icon {
-            wb = wb.with_window_icon(Icon::from_path(icon).ok());
-        }
-
-        self.winit_builder = wb;
+        self.config = config.clone();
+        self.winit_builder = config.to_windowbuilder(&self.events);
         self
     }